+++ CHANGELOG.md
@@ -2,6 +2,18 @@
 
 ## UNRELEASED
 
+* TeamsCallingPolicy
+  * Removed the AllowCalling parameter since it is no longer supported;
+  * Fixed an issue with Policies without tags in their name (e.g. Global);
+* ReverseDSC
+  * Change to allow ComponentsToExtract without the 'chck' prefix;
+* Metadata
+  * Updated MicrosoftTeams dependency to 1.0.5;
+  * Updated MSCloudLoginAssistant dependency to 1.0.3;
+  * Updated SharePointPnPPowerShellOnline dependency to 3.19.2003.0;
+
+## 1.0.3.1723
+
 * EXOHostedOutboundSpamFilterPolicy
   * BREAKING CHANGE: Remove IsSingleInstance and added
     Identity as key;